move example to next.js
