+++ packages/starter/example/.babelrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+    presets: [['next/babel']],
+    plugins: [['import', { libraryName: 'antd', style: true }]],
+};
