--- packages/starter/example/.prettierignore
@@ -1,1 +1,0 @@
-build
