+++ packages/starter/example/components/ContextProvider.tsx
@@ -1,0 +1,103 @@
+import { createTheme, ThemeProvider } from '@material-ui/core';
+import deepPurple from '@material-ui/core/colors/deepPurple';
+import pink from '@material-ui/core/colors/pink';
+import { WalletAdapterNetwork, WalletError } from '@solana/wallet-adapter-base';
+import { ConnectionProvider, WalletProvider } from '@solana/wallet-adapter-react';
+import {
+    getLedgerWallet,
+    getPhantomWallet,
+    getSlopeWallet,
+    getSolflareWallet,
+    getSolletExtensionWallet,
+    getSolletWallet,
+    getTorusWallet,
+} from '@solana/wallet-adapter-wallets';
+import { clusterApiUrl } from '@solana/web3.js';
+import { SnackbarProvider, useSnackbar } from 'notistack';
+import { FC, useCallback, useMemo } from 'react';
+import { AutoConnectProvider, useAutoConnect } from './AutoConnectProvider';
+
+const theme = createTheme({
+    palette: {
+        type: 'dark',
+        primary: {
+            main: deepPurple[700],
+        },
+        secondary: {
+            main: pink[700],
+        },
+    },
+    overrides: {
+        MuiButtonBase: {
+            root: {
+                justifyContent: 'flex-start',
+            },
+        },
+        MuiButton: {
+            root: {
+                textTransform: undefined,
+                padding: '12px 16px',
+            },
+            startIcon: {
+                marginRight: 8,
+            },
+            endIcon: {
+                marginLeft: 8,
+            },
+        },
+    },
+});
+
+const WalletContextProvider: FC = ({ children }) => {
+    const network = WalletAdapterNetwork.Devnet;
+    const endpoint = useMemo(() => clusterApiUrl(network), [network]);
+    const { autoConnect } = useAutoConnect();
+
+    const wallets = useMemo(
+        () => [
+            getPhantomWallet(),
+            getSlopeWallet(),
+            getSolflareWallet(),
+            getTorusWallet({
+                options: {
+                    clientId: 'BOM5Cl7PXgE9Ylq1Z1tqzhpydY0RVr8k90QQ85N7AKI5QGSrr9iDC-3rvmy0K_hF0JfpLMiXoDhta68JwcxS1LQ',
+                },
+            }),
+            getLedgerWallet(),
+            getSolletWallet({ network }),
+            getSolletExtensionWallet({ network }),
+        ],
+        [network]
+    );
+
+    const { enqueueSnackbar } = useSnackbar();
+    const onError = useCallback(
+        (error: WalletError) => {
+            enqueueSnackbar(error.message ? `${error.name}: ${error.message}` : error.name, { variant: 'error' });
+            console.error(error);
+        },
+        [enqueueSnackbar]
+    );
+
+    return (
+        <ConnectionProvider endpoint={endpoint}>
+            <WalletProvider wallets={wallets} onError={onError} autoConnect={autoConnect}>
+                {children}
+            </WalletProvider>
+        </ConnectionProvider>
+    );
+};
+
+const ContextProvider: FC = ({ children }) => {
+    return (
+        <ThemeProvider theme={theme}>
+            <SnackbarProvider>
+                <AutoConnectProvider>
+                    <WalletContextProvider>{children}</WalletContextProvider>
+                </AutoConnectProvider>
+            </SnackbarProvider>
+        </ThemeProvider>
+    );
+};
+
+export default ContextProvider;
