+++ packages/starter/example/components/notify.tsx
@@ -1,7 +1,7 @@
 import { Link, makeStyles } from '@material-ui/core';
 import LaunchIcon from '@material-ui/icons/Launch';
 import { useSnackbar, VariantType } from 'notistack';
-import React, { useCallback } from 'react';
+import { useCallback } from 'react';
 
 const useStyles = makeStyles({
     notification: {
