--- packages/starter/example/craco.config.js
@@ -1,17 +1,0 @@
-const CracoLessPlugin = require('craco-less');
-
-module.exports = {
-    plugins: [
-        {
-            plugin: CracoLessPlugin,
-            options: {
-                lessLoaderOptions: {
-                    lessOptions: {
-                        modifyVars: { '@primary-color': '#512da8' },
-                        javascriptEnabled: true,
-                    },
-                },
-            },
-        },
-    ],
-};
